requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.0
openpyxl>=3.0.0
pyinstaller>=5.0.0